requirements.txt
alibi-detect==0.11.4
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
apache-airflow==2.8.1
fastapi==0.110.1
uvicorn==0.29.0
joblib==1.4.0
pydantic==1.10.12
scipy==1.11.5
scikit-image==0.22.0
